Model Inference
This patch provides in-process execution of the model using tensorflow.

+++ preprocess.go
@@ -21,7 +21,7 @@ func NewTokenizer(alphabet string, maxChars int) Tokenizer {
 	return Tokenizer{dict, maxChars}
 }
 
-func (t Tokenizer) TextToSeq(text string) []int {
+func (t Tokenizer) TextToSeq(text string) []int64 {
 	runes := ([]rune)(strings.ToLower(text))
 
 	numChars := len(runes)
@@ -29,10 +29,10 @@ func (t Tokenizer) TextToSeq(text string) []int {
 		numChars = t.maxChars
 	}
 
-	seq := make([]int, t.maxChars)
+	seq := make([]int64, t.maxChars)
 	for i := numChars - 1; i >= 0; i-- {
 		no := t.dict[runes[i]]
-		seq[numChars-i-1] = no
+		seq[numChars-i-1] = int64(no)
 	}
 	return seq
 }
